+++ selfservice/hook/web_hook.go
@@ -81,6 +81,7 @@ type (
 		RequestURL     string             `json:"request_url"`
 		RequestCookies map[string]string  `json:"request_cookies"`
 		Identity       *identity.Identity `json:"identity,omitempty"`
+		Session        *session.Session   `json:"session,omitempty"`
 	}
 
 	WebHook struct {
@@ -137,7 +138,8 @@ func (e *WebHook) ExecuteLoginPostHook(_ http.ResponseWriter, req *http.Request,
 		RequestURL:     x.RequestURL(req).String(),
 		RequestCookies: cookies(req),
 		Identity:       session.Identity,
-	}, "LoginPostHook")
+		Session:        session,
+		}, "LoginPostHook")
 }
 
 func (e *WebHook) ExecuteVerificationPreHook(_ http.ResponseWriter, req *http.Request, flow *verification.Flow) error {